--- a/models/demos/yolov8x/runner/performant_runner_infra.py
+++ b/models/demos/yolov8x/runner/performant_runner_infra.py
@@ -47,15 +47,11 @@ class YOLOv8xPerformanceRunnerInfra:
             device=self.device, torch_model=torch_model, weights_mesh_mapper=self.weights_mesh_mapper
         )
 
-        # For multi-device scenarios, we use batch_size_per_device=1
-        # and replicate the same input across devices
-        batch_size_per_device = 1
-        input_shape = (batch_size_per_device, 3, 640, 640)
+        input_shape = (self.batch_size * self.num_devices, 3, 640, 640)
         self.torch_input_tensor = torch.randn(input_shape, dtype=torch.float32)
         self.tt_input_tensor = ttnn.from_torch(
             self.torch_input_tensor, ttnn.bfloat16, mesh_mapper=self.inputs_mesh_mapper
         )
-        # Compute expected output for the single device batch size
         self.torch_output_tensor = torch_model(self.torch_input_tensor)[0]
 
     def run(self):
@@ -84,25 +80,7 @@ class YOLOv8xPerformanceRunnerInfra:
 
         assert torch_input_tensor.ndim == 4, "Expected input tensor to have shape (BS, C, H, W)"
 
-        # For multi-device scenarios, create shards that match the mesh size
-        if device.get_num_devices() > 1:
-            # Create exactly the number of shards needed for the mesh
-            num_shards_needed = device.get_num_devices()
-
-            # Split the input tensor into the required number of shards
-            # For batch size 1 with 2 devices, we need to replicate the same input for both devices
-            input_tensor = []
-            for i in range(num_shards_needed):
-                # Use the same input tensor for all shards (replication)
-                input_tensor.append(torch_input_tensor)
-
-            tt_inputs_host = ttnn.from_host_shards(
-                [ttnn.from_torch(t, dtype=ttnn.bfloat16, layout=ttnn.ROW_MAJOR_LAYOUT) for t in input_tensor],
-                device.shape,
-            )
-        else:
-            # Single device scenario
-            tt_inputs_host = ttnn.from_torch(torch_input_tensor, ttnn.bfloat16, mesh_mapper=self.inputs_mesh_mapper)
+        tt_inputs_host = ttnn.from_torch(torch_input_tensor, ttnn.bfloat16, mesh_mapper=self.inputs_mesh_mapper)
 
         return tt_inputs_host, input_mem_config
 
@@ -127,21 +105,12 @@ class YOLOv8xPerformanceRunnerInfra:
 
     def validate(self, output_tensor=None):
         if output_tensor is None:
-            # Use self.output_tensor (original behavior)
             output_tensor = self.output_tensor
         else:
-            # Handle output from pipeline (which might be a list or multi-device tensor)
             if isinstance(output_tensor, (list, tuple)):
-                # Extract the first element from the list
                 output_tensor = output_tensor[0]
 
         output_tensor = ttnn.to_torch(output_tensor, mesh_composer=self.outputs_mesh_composer)
-
-        # For multi-device scenarios, we only validate against the first device's output
-        # since all devices should produce identical results with the same input
-        if self.num_devices > 1 and output_tensor.shape[0] > 1:
-            # Take only the first device's output for validation
-            output_tensor = output_tensor[0:1]  # Keep the batch dimension
 
         valid_pcc = 0.978
         self.pcc_passed, self.pcc_message = assert_with_pcc(self.torch_output_tensor, output_tensor, pcc=valid_pcc)