### Summary
This PR refactors the performance tests for YOLO models (YoloV4, YoloV7, and YoloV8x) to use the TT-CNN pipeline API.

- YoloV4: Refactored `models/demos/yolov4/tests/perf/test_e2e_performant.py` to use TT-CNN pipeline with `create_pipeline_from_config()` and `preallocate_output_tensors_on_host()`; 
    - Performance at 320x320 improved from 109 → 130 FPS
    - Performance at 640x640 improved from 48 → 65 FPS
- YoloV4: Updated demo script to use TT-CNN pipeline
- YoloV7: Updated `models/demos/yolov7/tests/perf/test_e2e_performant.py` to use TT-CNN pipeline
- YoloV8x: Refactored `models/demos/yolov8x/tests/perf/test_e2e_performant.py` to use TT-CNN pipeline
- Updated `models/demos/ttnn_resnet/tests/perf_e2e_resnet50.py `to use output tensor preallocation using `preallocate_output_tensors_on_host()` to reduce memory allocation overhead during inference
- Removed the need to specify output tensor shape/dtype/layout when using preallocated host tensors
- Added support for models with multiple outputs in the TT-CNN pipeline framework
- Updated executor implementations to handle structured outputs (lists/tuples of tensors)

### Future Work
- Update demo and web demo code to use TT-CNN pipeline API so that we can remove the trace+2CQ implementations for each YoloV4/YoloV7/YoloV8x
   
### Checklist
- [x] [All post commit](https://github.com/tenstorrent/tt-metal/actions/workflows/all-post-commit-workflows.yaml) CI passes: https://github.com/tenstorrent/tt-metal/actions/runs/17245014334
- [x] [Blackhole Post commit](https://github.com/tenstorrent/tt-metal/actions/workflows/blackhole-post-commit.yaml) CI with demo tests passes (if applicable): https://github.com/tenstorrent/tt-metal/actions/runs/17245023027
- [x] [Model regression](https://github.com/tenstorrent/tt-metal/actions/workflows/perf-models.yaml) CI passes (if applicable): https://github.com/tenstorrent/tt-metal/actions/runs/17295374211
- [x] [Model regression (T3K)](https://github.com/tenstorrent/tt-metal/actions/workflows/perf-models.yaml) CI passes (if applicable): https://github.com/tenstorrent/tt-metal/actions/runs/17237134779/job/48905250102
- [ ] [Demo tests (single card)](https://github.com/tenstorrent/tt-metal/actions/workflows/perf-models.yaml) CI passes (if applicable): https://github.com/tenstorrent/tt-metal/actions/runs/17296665803