@pavlejosipovic 2025-08-27
Any chance we can point claude code to this commit and make it refactor the resto of yolov models?